# docs/evidence/contracts/iskai-openapi.current.json
{
  "components": {
    "schemas": {
      "Body_extract_metadata_upload_extract_metadata_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_extract_metadata_upload_extract_metadata_post",
        "type": "object"
      },
...
      "Body_upload_paper_upload_paper_post": {
        "properties": {
          "abstract": {
            "default": "",
            "title": "Abstract",
            "type": "string"
          },
          "authors": {
            "default": "",
            "title": "Authors",
            "type": "string"
          },
          "department": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "program_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Program Id"
          },
          "specialization_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialization Id"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "track": {
            "default": "",
            "title": "Track",
            "type": "string"
          },
          "year": {
            "default": "",
            "title": "Year",
            "type": "string"
          }
        },
        "required": [
          "file",
          "title"
        ],
        "title": "Body_upload_paper_upload_paper_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UploadAccepted": {
        "properties": {
          "idempotency_key": {
            "title": "Idempotency Key",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "idempotency_key",
          "status",
          "message"
        ],
        "title": "UploadAccepted",
        "type": "object"
      },
      "UploadCancelRequest": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "UploadCancelRequest",
        "type": "object"
      },
      "UploadCancelResponse": {
        "properties": {
          "cancel_requested": {
            "default": false,
            "title": "Cancel Requested",
            "type": "boolean"
          },
          "cancelled_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "outcome": {
            "title": "Outcome",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "outcome",
          "status",
          "message"
        ],
        "title": "UploadCancelResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Centralized AI-Powered Thesis Library using Retrieval-Augmented Generation for the College of Computing Studies, Information and Communication Technology, Isabela State University, Echague.",
    "title": "ISU Thesis AI Library API",
    "version": "2.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/upload/extract-metadata": {
      "post": {
        "description": "Extract thesis metadata locally, with Gemini filling missing fields.",
        "operationId": "extract_metadata_upload_extract_metadata_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_extract_metadata_upload_extract_metadata_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "The request payload or parameters were rejected."
          },
          "413": {
            "description": "The uploaded file exceeds the configured size limit."
          },
          "415": {
            "description": "The uploaded file type is not accepted."
          },
          "422": {
            "description": "The request was well formed but failed validation."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Extract Metadata",
        "tags": [
          "upload"
        ]
      }
    },
    "/upload/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_upload_job_upload_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadCancelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "The authenticated account is not permitted to perform this action."
          },
          "404": {
            "description": "The resource does not exist, or is not visible to this account."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "A required backing service is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Upload Job",
        "tags": [
          "upload"
        ]
      }
    },
    "/upload/paper": {
      "post": {
        "operationId": "upload_paper_upload_paper_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_paper_upload_paper_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "The request payload or parameters were rejected."
          },
          "409": {
            "description": "The request conflicts with the current state of the resource."
          },
          "413": {
            "description": "The uploaded file exceeds the configured size limit."
          },
          "415": {
            "description": "The uploaded file type is not accepted."
          },
          "422": {
            "description": "The request was well formed but failed validation."
          },
          "503": {
            "description": "A required backing service is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Paper",
        "tags": [
          "upload"
        ]
      }
    },
    "/upload/status/{job_id}": {
      "get": {
        "operationId": "upload_status_upload_status__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadJobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "The resource does not exist, or is not visible to this account."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "A required backing service is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Status",
        "tags": [
          "upload"
        ]
      }
    },
...
  }
}